+++ backends/xla/builder.go
@@ -63,7 +63,7 @@ func (b *Builder) DistributedSPMD(numDevices int) error {
 
 // DistributedAutoSharding is not supported by the old XLA backend.
 func (b *Builder) DistributedAutoSharding(meshes ...backends.Mesh) error {
-	return errors.Wrapf(notimplemented.NotImplementedError, "in DistributedAutoSharding()")
+	return errors.Wrapf(notimplemented.NotImplementedError, "in Builder.DistributedAutoSharding()")
 }
 
 // DeviceAssignment assigns the devices to the computation.
@@ -111,12 +111,22 @@ func (b *Builder) verifyAndCastOp(op backends.Op, paramName string) (*xlabuilder
 	}
 	xlaOp, ok := op.(*xlabuilder.Op)
 	if !ok {
-		return nil, errors.Errorf("nil or invalid Op (%T: %v) given as an input %s, it must be an Op created by the same backend builder (%s:%s)",
-			op, op, paramName, b.backend.Name(), b.name)
+		return nil, errors.Errorf(
+			"nil or invalid Op (%T: %v) given as an input %s, it must be an Op created by the same backend builder (%s:%s)",
+			op,
+			op,
+			paramName,
+			b.backend.Name(),
+			b.name,
+		)
 	}
 	if xlaOp.Builder() != b.builder {
-		return nil, errors.Errorf("op given to parameter %s was created with a different builder (%s) than the builder (%s) it is being used in -- Ops cannot cross to different builders",
-			paramName, xlaOp.Builder().Name(), b.Name())
+		return nil, errors.Errorf(
+			"op given to parameter %s was created with a different builder (%s) than the builder (%s) it is being used in -- Ops cannot cross to different builders",
+			paramName,
+			xlaOp.Builder().Name(),
+			b.Name(),
+		)
 	}
 	return xlaOp, nil
 }
@@ -148,7 +158,12 @@ func (b *Builder) OpShape(op backends.Op) (shapes.Shape, error) {
 
 // Parameter creates an input parameter for the computation.
 // During execution of the computation this value will need to be fed, in the same order it is created.
-func (b *Builder) Parameter(name string, shape shapes.Shape) (backends.Op, error) {
+func (b *Builder) Parameter(name string, shape shapes.Shape, spec backends.ShardingSpec) (backends.Op, error) {
+	if len(spec) > 0 {
+		return nil, errors.Wrapf(
+			notimplemented.NotImplementedError,
+			"sharding spec %v not supported for %q builder", spec, BackendName)
+	}
 	op, err := xlabuilder.Parameter(b.builder, name, len(b.parameterNames), shapeToXShape(shape))
 	if err != nil {
 		return nil, errors.WithMessagef(err, "backend %q: Parameter(%q, %s)", BackendName, name, shape)
@@ -178,11 +193,25 @@ func (b *Builder) Constant(flat any, dims ...int) (backends.Op, error) {
 	}
 	literal, err := xlabuilder.NewArrayLiteralFromAny(flat, dims...)
 	if err != nil {
-		return nil, errors.WithMessagef(err, "backend %q builder %q: Constant(%T, dims=%v)", b.backend.Name(), b.name, flat, dims)
+		return nil, errors.WithMessagef(
+			err,
+			"backend %q builder %q: Constant(%T, dims=%v)",
+			b.backend.Name(),
+			b.name,
+			flat,
+			dims,
+		)
 	}
 	op, err := xlabuilder.Constant(b.builder, literal)
 	if err != nil {
-		return nil, errors.WithMessagef(err, "backend %q builder %q: Constant(%T, dims=%v)", b.backend.Name(), b.name, flat, dims)
+		return nil, errors.WithMessagef(
+			err,
+			"backend %q builder %q: Constant(%T, dims=%v)",
+			b.backend.Name(),
+			b.name,
+			flat,
+			dims,
+		)
 	}
 	return op, nil
 }
@@ -197,7 +226,12 @@ func (b *Builder) Identity(x backends.Op) (backends.Op, error) {
 	return xlabuilder.Identity(xlaX), nil
 }
 
-func (b *Builder) ReduceWindow(x backends.Op, reductionType backends.ReduceOpType, windowDimensions, strides, baseDilations, windowDilations []int, paddings [][2]int) (backends.Op, error) {
+func (b *Builder) ReduceWindow(
+	x backends.Op,
+	reductionType backends.ReduceOpType,
+	windowDimensions, strides, baseDilations, windowDilations []int,
+	paddings [][2]int,
+) (backends.Op, error) {
 	xlaX, err := b.verifyAndCastOp(x, "x")
 	if err != nil {
 		return nil, err
@@ -217,11 +251,21 @@ func (b *Builder) ReduceWindow(x backends.Op, reductionType backends.ReduceOpTyp
 	case backends.ReduceOpProduct:
 		cfg = cfg.Product()
 	default:
-		return nil, errors.Errorf("unknown reduction type %s given to ReduceWindow (building %q)", reductionType, b.name)
+		return nil, errors.Errorf(
+			"unknown reduction type %s given to ReduceWindow (building %q)",
+			reductionType,
+			b.name,
+		)
 	}
 	op, err := cfg.Done()
 	if err != nil {
-		return nil, errors.WithMessagef(err, "backend %q builder %q: ReduceWindow(reductionType=%s)", b.backend.Name(), b.name, reductionType)
+		return nil, errors.WithMessagef(
+			err,
+			"backend %q builder %q: ReduceWindow(reductionType=%s)",
+			b.backend.Name(),
+			b.name,
+			reductionType,
+		)
 	}
 	return op, nil
 }
@@ -239,7 +283,13 @@ func (b *Builder) RngBitGenerator(state backends.Op, shape shapes.Shape) (newSta
 	xlaShape := shapeToXShape(shape)
 	newState, values, err = xlabuilder.RngBitGenerator(xlaState, xlaShape)
 	if err != nil {
-		return nil, nil, errors.WithMessagef(err, "backend %q builder %q: RngBitGenerator(shape=%s)", b.backend.Name(), b.name, shape)
+		return nil, nil, errors.WithMessagef(
+			err,
+			"backend %q builder %q: RngBitGenerator(shape=%s)",
+			b.backend.Name(),
+			b.name,
+			shape,
+		)
 	}
 	return
 }
@@ -251,7 +301,11 @@ func (b *Builder) RngBitGenerator(state backends.Op, shape shapes.Shape) (newSta
 //
 // Based on paper "Batch Normalization: Accelerating Deep Network Training by Reducing
 // Internal Covariate Shift" (Sergey Ioffe, Christian Szegedy), https://arxiv.org/abs/1502.03167.
-func (b *Builder) BatchNormForTraining(operand, scale, offset backends.Op, epsilon float32, axis int) (normalized, batchMean, batchVariance backends.Op, err error) {
+func (b *Builder) BatchNormForTraining(
+	operand, scale, offset backends.Op,
+	epsilon float32,
+	axis int,
+) (normalized, batchMean, batchVariance backends.Op, err error) {
 	xlaOperand, err := b.verifyAndCastOp(operand, "operand")
 	if err != nil {
 		return
@@ -264,9 +318,21 @@ func (b *Builder) BatchNormForTraining(operand, scale, offset backends.Op, epsil
 	if err != nil {
 		return
 	}
-	normalized, batchMean, batchVariance, err = xlabuilder.BatchNormForTraining(xlaOperand, xlaScale, xlaOffset, epsilon, axis)
+	normalized, batchMean, batchVariance, err = xlabuilder.BatchNormForTraining(
+		xlaOperand,
+		xlaScale,
+		xlaOffset,
+		epsilon,
+		axis,
+	)
 	if err != nil {
-		err = errors.WithMessagef(err, "backend %q builder %q: BatchNormForTraining(operand=%s)", b.backend.Name(), b.name, xlaOperand.Shape)
+		err = errors.WithMessagef(
+			err,
+			"backend %q builder %q: BatchNormForTraining(operand=%s)",
+			b.backend.Name(),
+			b.name,
+			xlaOperand.Shape,
+		)
 		return
 	}
 	return
@@ -282,7 +348,11 @@ func (b *Builder) BatchNormForTraining(operand, scale, offset backends.Op, epsil
 //
 // Based on paper "Batch Normalization: Accelerating Deep Network Training by Reducing
 // Internal Covariate Shift" (Sergey Ioffe, Christian Szegedy), https://arxiv.org/abs/1502.03167.
-func (b *Builder) BatchNormGradient(operand, scale, mean, variance, gradOutput backends.Op, epsilon float32, axis int) (gradOperand, gradScale, gradOffset backends.Op, err error) {
+func (b *Builder) BatchNormGradient(
+	operand, scale, mean, variance, gradOutput backends.Op,
+	epsilon float32,
+	axis int,
+) (gradOperand, gradScale, gradOffset backends.Op, err error) {
 	xlaOperand, err := b.verifyAndCastOp(operand, "operand")
 	if err != nil {
 		err = errors.WithMessagef(err, "calling BatchNormGradient()")
@@ -308,9 +378,23 @@ func (b *Builder) BatchNormGradient(operand, scale, mean, variance, gradOutput b
 		err = errors.WithMessagef(err, "calling BatchNormGradient()")
 		return
 	}
-	gradOperand, gradScale, gradOffset, err = xlabuilder.BatchNormGradient(xlaOperand, xlaScale, xlaMean, xlaVariance, xlaGradOutput, epsilon, axis)
+	gradOperand, gradScale, gradOffset, err = xlabuilder.BatchNormGradient(
+		xlaOperand,
+		xlaScale,
+		xlaMean,
+		xlaVariance,
+		xlaGradOutput,
+		epsilon,
+		axis,
+	)
 	if err != nil {
-		err = errors.WithMessagef(err, "backend %q builder %q: BatchNormGradient(operand=%s)", b.backend.Name(), b.name, xlaOperand.Shape)
+		err = errors.WithMessagef(
+			err,
+			"backend %q builder %q: BatchNormGradient(operand=%s)",
+			b.backend.Name(),
+			b.name,
+			xlaOperand.Shape,
+		)
 		return
 	}
 	return